change text styles type from int to double

+++ lib/constants.dart
@@ -7,10 +7,10 @@ const mainText = Color.fromRGBO(41, 62, 93, 1);
 const secondaryText = Color.fromRGBO(159, 165, 192, 1);
 
 // Text styles
-const int headingOne = 22;
-const int headingTwo = 17;
-const int headingThree = 15;
+const double headingOne = 22;
+const double headingTwo = 17;
+const double headingThree = 15;
 
-const int paragraphOne = 17;
-const int paragraphTwo = 15;
-const int smallText = 12;
+const double paragraphOne = 17;
+const double paragraphTwo = 15;
+const double smallText = 12;
